Minicute.Data.Minicute.Expression

Description

Types for a miniCUTE expression

Documentation

module Minicute.Data.Common

newtype LetDefinition t l a

A type used to represent a let definition of expr a.

Constructors

LetDefinition (a, Expression t l a)

type MainLetDefinition t l = LetDefinition t l Identifier

A MainLetDefinition for ExpressionMC with Identifier.

_letDefinitionBinder :: Lens' (LetDefinition t l a) a

Lens to extract the binder of LetDefinition

_letDefinitionBody :: Lens (LetDefinition t l a) (LetDefinition t' l' a) (Expression t l a) (Expression t' l' a)

Lens to extract the body expression of LetDefinition

newtype MatchCase t l a

A type used to represent a match case of expr a.

Constructors

MatchCase (Integer, [a], Expression t l a)

type MainMatchCase t l = MatchCase t l Identifier

A MatchCase for ExpressionMC with Identifier.

_matchCaseTag :: Lens' (MatchCase t l a) Integer

Lens to extract the tag of MatchCase

_matchCaseArguments :: Lens' (MatchCase t l a) [a]

Lens to extract the arguments of MatchCase

_matchCaseBody :: Lens (MatchCase t l a) (MatchCase t' l' a) (Expression t l a) (Expression t' l' a)

Lens to extract the body expression of MatchCase

data IsRecursive

IsRecursive represents recursiveness of let/letrec expressions.

pattern Recursive :: IsRecursive

Utility pattern for the recursive case of IsRecursive

pattern NonRecursive :: IsRecursive

Utility pattern for the non-recursive case of IsRecursive

data ExpressionType

Constructors

Simple

AnnotatedWith Type

data ExpressionLevel

Constructors

MC
miniCUTE

LLMC
Lambda lifted miniCUTE

data Expression (t :: ExpressionType) (l :: ExpressionLevel) a where

A basic miniCUTE expression of a.

Constructors

EInteger :: Annotation t -> Integer -> Expression t l a
5

EConstructor :: Annotation t -> Integer -> Integer -> Expression t l a
$C{t;a}

EPrimitive :: Annotation t -> Primitive -> Expression t l a
+

EVariable :: Annotation t -> Identifier -> Expression t l a
v

EApplication :: Annotation t -> Expression t l a -> Expression t l a -> Expression t l a
f 4

ELet :: Annotation t -> IsRecursive -> [LetDefinition t l a] -> Expression t l a -> Expression t l a
let x = 4 in x

EMatch :: Annotation t -> Expression t l a -> [MatchCase t l a] -> Expression t l a
match $C{1;0} with <1> -> 4

ELambda :: l ~ MC => Annotation t -> [a] -> Expression t l a -> Expression t l a
\x.x

type MainExpression t l = Expression t l Identifier

A Expression with Identifier.

pattern EApplication2 :: forall (t :: ExpressionType) (l :: ExpressionLevel) a. Annotation t -> Annotation t -> Expression t l a -> Expression t l a -> Expression t l a -> Expression t l a

A utility pattern for Expression of double application.

pattern EApplication3 :: forall (t :: ExpressionType) (l :: ExpressionLevel) a. Annotation t -> Annotation t -> Annotation t -> Expression t l a -> Expression t l a -> Expression t l a -> Expression t l a -> Expression t l a

A utility pattern for Expression of triple application.

_annotation :: Lens' (Expression t l a) (Annotation t)

Lens to extract the annotation of AnnotatedExpressionMC.